refactor(backend): refactored admin transaction to be in repository (SCRUM-55)
