+++ backend/src/modules/users/users.repository.ts
@@ -209,7 +209,6 @@ export class UsersRepository implements IUsersRepository {
           tag_id: tag.tag_id,
           priority: tag.priority,
         })),
-        skipDuplicates: false,
       });
       operations.push(createTagsOperation);
     }
